requirements.txt
asgiref==3.4.0
autopep8==1.5.7
beautifulsoup4==4.9.3
Django==3.2.4
djangorestframework==3.12.4
psycopg2==2.8.6
pycodestyle==2.7.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
yahoofinancials==1.6
